* Create student validator, CRD operations

* feat(blocknote): blocknote

ni

* feat(created the blocknote module): students and teachers can add task

Task can be added and set deadline, students and teachers can also add notes

* feat(added deadline to the schema): deadline was added

Deadline can be added on tasks

* Leaves, TimeTable and Payroll tasks

* Student CRUD operation, removal of 'password' from Student model

* Add student routes

* HRM with validators and Swager doc

* Added student route, corrected Gender enum for student validator

* refactor(human resources): refactor the HR module and generate student's swagger file

refactor the HR components into a single module, generate the swagger file of the student module

* refactor(module): structure the academic

refactor the academics components into a single module, resolved merged conflics, add the swagger
file for the academic year

* fix(module): merge conflic, code structure

resolved the merge conflic, update the routes, and add the swagger configuration for the blocknote

* feat(validation): inputs are now validated

inputs are validated so that only the correct data is sent

* Add param to delete and update route

* feat(file): Incident module implement with crud ops (#16)

feat(file): added the crud operations  on the incident modules

* Updated student controller and route file

* testing

* perf(component): improve the database and the admin component

adjust the prisma schema accordingly to let the database beign consistant, improve the routes of the
incident and school event componenets by updating the routes and the swagger file

* fix(academics module): adjust the fields accordingly with respect to the database

adjust the fields by deleting the class model and maintaining the classes model because they where
refering the same model, avoiding keywords naming like that of "class" and adjust the controller
academics to match with the prisma schema

* perf(modules): update the academics, student, and blocknote module

academics, update the data validation, update the routes and add in the authentication middleware,
improve the logic in the controller and the swagger file configuration. Update the student
validations, authentication and add in the endpoint to get a student by id. Add the authentication
middleware in the blocknote to let authenticated users take notes in the system

* perf(module): the human resource code improved

improve the routes and put in the authentication of users, adjust the controller accordingly and
update the swagger file of the academic module

---------

+++ src/controllers/student.controller.ts
@@ -1,12 +1,15 @@
 import { Request, Response } from 'express';
 import { PrismaClient } from '@prisma/client';
-import { CreateStudentDto } from '../valdators/student.validator';
+import {
+  CreateStudentDto,
+  UpdateStudentDto,
+} from '../valdators/student.validator';
 import { validate } from 'class-validator';
 
 const prisma = new PrismaClient();
 
 /*
-  @route    POST: /students
+  @route    POST: /students/create
   @access   private
   @desc     Create a new student
 */
@@ -22,7 +25,7 @@ export const createStudent = async (req: Request, res: Response) => {
     if (errors.length > 0) {
       return res.status(400).json({
         type: 'error',
-        message: 'Validation failed',
+        message: 'Invalid details',
         errors: errors,
       });
     }
@@ -31,11 +34,13 @@ export const createStudent = async (req: Request, res: Response) => {
       data: {
         First_Name: studentData.First_Name,
         Last_Name: studentData.Last_Name,
-        Date_Of_Birth: studentData.Date_Of_Birth, //2024-09-24T00:00:00.000Z
+        Date_Of_Birth: new Date(studentData.Date_Of_Birth),
         Address: studentData.Address,
         Gender: studentData.Gender,
         Image: studentData.Image,
         Phone: studentData.Phone,
+        classId: studentData.classId,
+        Medical_Info: studentData.Medical_Info,
       },
     });
 
@@ -54,13 +59,18 @@ export const createStudent = async (req: Request, res: Response) => {
   }
 };
 
-export const getStudent = async (req: Request, res: Response) => {
+/*
+  @route    GET: /students
+  @access   private
+  @desc     Retrieve all students
+*/
+export const getStudents = async (req: Request, res: Response) => {
   try {
     const students = await prisma.student.findMany();
 
-    res.json({
+    res.status(200).json({
       type: 'success',
-      message: '',
+      message: 'Successfully retrieved students',
       data: students,
     });
   } catch (error) {
@@ -73,15 +83,69 @@ export const getStudent = async (req: Request, res: Response) => {
   }
 };
 
+/*
+  @route    GET: /students/:id
+  @access   private
+  @desc     Retrieve a student by ID
+*/
+export const getStudentById = async (req: Request, res: Response) => {
+  const { id } = req.params;
+
+  if (isNaN(Number(id))) {
+    return res.status(404).json({
+      type: 'error',
+      message: 'Invalid student ID',
+    });
+  }
+
+  try {
+    const student = await prisma.student.findUnique({
+      where: { Student_ID: Number(id) },
+    });
+
+    if (!student) {
+      return res.status(404).json({
+        type: 'error',
+        message: 'Student not found',
+      });
+    }
+
+    res.status(200).json({
+      type: 'success',
+      message: 'Successfully retrieved student',
+      data: student,
+    });
+  } catch (error) {
+    console.error('Error retrieving student:', error);
+    res.status(500).json({
+      type: 'error',
+      message: 'Error retrieving student',
+      data: {},
+    });
+  }
+};
+
+/*
+  @route    PUT: /students/update/:id
+  @access   private
+  @desc     Update a student's information
+*/
 export const updateStudent = async (req: Request, res: Response) => {
   const { id } = req.params;
 
+  if (isNaN(Number(id))) {
+    return res.status(404).json({
+      type: 'error',
+      message: 'Invalid student ID',
+    });
+  }
+
   try {
     const updateData = req.body;
-    const createStudentDto = new CreateStudentDto();
-    Object.assign(createStudentDto, updateData); // Assign the incoming data to the DTO
+    const updateStudentDto = new UpdateStudentDto();
+    Object.assign(updateStudentDto, updateData);
 
-    const errors = await validate(createStudentDto);
+    const errors = await validate(updateStudentDto);
     if (errors.length > 0) {
       return res.status(400).json({
         type: 'error',
@@ -90,12 +154,12 @@ export const updateStudent = async (req: Request, res: Response) => {
       });
     }
 
-    const updatedstudent = await prisma.student.update({
+    const updatedStudent = await prisma.student.update({
       where: { Student_ID: Number(id) },
       data: {
         First_Name: updateData.First_Name,
         Last_Name: updateData.Last_Name,
-        Date_Of_Birth: updateData.Date_Of_Birth,
+        Date_Of_Birth: new Date(updateData.Date_Of_Birth),
         Gender: updateData.Gender,
         Address: updateData.Address,
         Phone: updateData.Phone,
@@ -103,10 +167,10 @@ export const updateStudent = async (req: Request, res: Response) => {
       },
     });
 
-    res.json({
+    res.status(200).json({
       type: 'success',
-      message: `student with ID ${id} updated successfully`,
-      data: updatedstudent,
+      message: `Student with ID ${id} updated successfully`,
+      data: updatedStudent,
     });
   } catch (error) {
     console.error('Error updating student:', error);
@@ -118,15 +182,27 @@ export const updateStudent = async (req: Request, res: Response) => {
   }
 };
 
+/*
+  @route    DELETE: /students/delete/:id
+  @access   private
+  @desc     Delete a student
+*/
 export const deleteStudent = async (req: Request, res: Response) => {
   const { id } = req.params;
+
+  if (isNaN(Number(id))) {
+    return res.status(404).json({
+      type: 'error',
+      message: 'Invalid student ID',
+    });
+  }
 
   try {
     await prisma.student.delete({
       where: { Student_ID: Number(id) },
     });
 
-    res.json({
+    res.status(200).json({
       type: 'success',
       message: `Student with ID ${id} deleted successfully`,
       data: {},
